requirements.txt
instagrapi
pillow
js2py